feat(geography)!: move error matcher injection token to state subpackage (#2645)
BREAKING CHANGE: error matcher injection token has moved to state subpackage
