+++ libs/geography/src/public_api.ts
@@ -3,4 +3,3 @@
  */
 export * from './models/public_api';
 export * from './comparators/public_api';
-export * from './injection-tokens/public_api';
