+++ libs/geography/state/src/index.ts
@@ -1,7 +1,1 @@
-export * from './selectors/public_api';
-export * from './reducers/public_api';
-export * from './actions/public_api';
-
-export { DaffGeographyFacade } from './facades/geography/geography.facade';
-export { DaffGeographyFacadeInterface } from './facades/geography/geography-facade.interface';
-export { DaffGeographyStateModule } from './geography-state.module';
+export * from './public_api';
